Added xEnv to Sol
This allows us to get the (BindId, Sort) for any symbols in the entire
SInfo. It's needed so that we can fold solveEbinds into lhsPred without
having to recompute it every time.

We could probably fold this into Sol's sEnv field, and should try to do
so after we finish implementing the solveEbinds feature.