Get rid of the FinalExpand flag in the PLE implementation
This flag was an optimization to avoid comparing expressions.
It was only read by REST which has more serious bottlenecks to worry about,
and the exact meaning of the flag was unclear. It was not equivalent to
comparing the expressions, and finding the difference required
reading a lot of code.